fix: split out the classes in describers.ts
