+++ packages/salesforcedx-vscode-core/src/commands/forceSourceRetrieveMetadata/index.ts
@@ -4,7 +4,7 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-export { RetrieveDescriberFactory } from './describers';
 export { forceSourceRetrieveCmp } from './forceSourceRetrieveCmp';
 export { RetrieveDescriber } from './retrieveDescriber';
+export { RetrieveDescriberFactory } from './retrieveDescriberFactory';
 export { RetrieveMetadataTrigger } from './retrieveMetadataTrigger';
